LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/security.pattern.fillStyle.sub-expected.txt

Blocked access to external URL http://www1.localhost:8800/images/red.png
Blocked access to external URL http://www1.localhost:8800/images/red.png
Blocked access to external URL http://www1.localhost:8800/media/movie_300.ogv
Blocked access to external URL http://www1.localhost:8800/media/movie_300.ogv
Setting fillStyle to a pattern of an unclean canvas makes the canvas origin-unclean


Harness Error (TIMEOUT), message = null

FAIL cross-origin HTMLImageElement: Setting fillStyle to an origin-unclear pattern makes the canvas origin-unclean promise_test: Unhandled rejection with value: object "[object Event]"
FAIL cross-origin SVGImageElement: Setting fillStyle to an origin-unclear pattern makes the canvas origin-unclean promise_test: Unhandled rejection with value: object "[object Event]"
FAIL cross-origin HTMLVideoElement: Setting fillStyle to an origin-unclear pattern makes the canvas origin-unclean promise_test: Unhandled rejection with value: object "[object Event]"
TIMEOUT redirected to cross-origin HTMLVideoElement: Setting fillStyle to an origin-unclear pattern makes the canvas origin-unclean Test timed out
NOTRUN redirected to same-origin HTMLVideoElement: Setting fillStyle to an origin-unclear pattern makes the canvas origin-unclean 
NOTRUN unclean HTMLCanvasElement: Setting fillStyle to an origin-unclear pattern makes the canvas origin-unclean 
NOTRUN unclean ImageBitmap: Setting fillStyle to an origin-unclear pattern makes the canvas origin-unclean 
